Create css-major.html
Use your heading tags for headings and paragraph tags for the paragraphs. Also format the write-ups with paddings, margins, and borders to beautify them.

The entire write-up should have a background color “white” and it should be in three sections


The first section: Write the heading “First section” and comment it out (use 18px). Let it contain your name and your love for technology. The text color should be green. Give formatting to it using bold tags and italics tags and the size should be 14px.


The second section: Write the heading “Second section” and comment it out (use 18px). Let it contain what you have learned so far and how css is important. The text color should be red. Give formatting to it using bold tags and italics tags and the size should be 14px


The third section: Write the heading “Third section” and comment it out (use 18px). Explaining the difference between paddings, margins, and borders. The text color should be blue. Give formatting to it using bold tags and italics tags and the size should be 14px.